+++ src/components/MapViewManager.tsx
@@ -11,13 +11,9 @@ import {
  * Internal dependencies
  */
 import type { HardwareKeyEventResponse, Location, ResponseInclude } from '../types';
+import { LINKING_ERROR } from '../constants';
 
 const ComponentName = 'MapViewManager';
-
-const LINKING_ERROR =
-    `The package 'react-native-mapsforge-vtm' doesn't seem to be linked. Make sure: \n\n` +
-    '- You rebuilt the app after installing the package\n' +
-    '- You are not using Expo Go\n';
 
 type Props = {
 	ref: MutableRefObject<number | Component<any, any, any> | ComponentClass<any, any> | null>;
@@ -27,9 +23,9 @@ type Props = {
 	heightForLayoutSize: number;
 	center: Location;
 	moveEnabled: 1 | 0;
-	tiltEnabled: 1 | 0
-	rotationEnabled: 1 | 0
-	zoomEnabled: 1 | 0
+	tiltEnabled: 1 | 0;
+	rotationEnabled: 1 | 0;
+	zoomEnabled: 1 | 0;
 	zoomLevel: number;
 	zoomMin: number;
 	zoomMax: number;
@@ -44,8 +40,10 @@ type Props = {
 	maxRoll: number;
 	hgtDirPath?: `/${string}` | `content://${string}`;
 	responseInclude: ResponseInclude;
+	hgtInterpolation:  1 | 0;
 	mapEventRate: number;
 	hgtReadFileRate: number;
+	hgtFileInfoPurgeThreshold: number;
 	emitsMapEvents: 1 | 0;
 	emitsHardwareKeyUp: HardwareKeyEventResponse['keyCodeString'][];
 };